+++ packages/components/skeleton/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @jala-banyu/skeleton
 
+## 3.0.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.2.0
+  - @jala-banyu/system-rsc@3.0.0
+
 ## 2.0.1
 
 ### Patch Changes